Add files via upload
This games_subset.csv is from the RAWG API I used for a data project. It shows a list of games released in the last year with the rating, review count, genre, and ESRB rating among other features. Potentially, it could be used to predict future ratings for similar titles based on the features.